requirements.txt
numpy==1.18.2
pytorch==1.7.0
scikit-learn==0.23.2
scipy==1.5.2